+++ bin/bench-transaction/src/note_costs.rs
@@ -45,7 +45,7 @@ impl PricedNote {
             PricedNote::Standard(StandardNote::BURN),
             PricedNote::Standard(StandardNote::CONSTANT_FEE_POLICY_CONFIG),
             PricedNote::Standard(StandardNote::FAUCET_POLICY_CONFIG),
-            PricedNote::Standard(StandardNote::FUNGIBLE_FAUCET_CONFIG),
+            PricedNote::Standard(StandardNote::FAUCET_METADATA_CONFIG),
             PricedNote::Standard(StandardNote::ALLOWLIST_CONFIG),
             PricedNote::Standard(StandardNote::BLOCKLIST_CONFIG),
             PricedNote::Standard(StandardNote::PAUSE_CONFIG),
@@ -96,8 +96,8 @@ impl PricedNote {
             PricedNote::Standard(StandardNote::FAUCET_POLICY_CONFIG) => {
                 &[ExecutionBenchmark::ConsumeFaucetPolicyConfigNetwork]
             },
-            PricedNote::Standard(StandardNote::FUNGIBLE_FAUCET_CONFIG) => {
-                &[ExecutionBenchmark::ConsumeFungibleFaucetConfigNetwork]
+            PricedNote::Standard(StandardNote::FAUCET_METADATA_CONFIG) => {
+                &[ExecutionBenchmark::ConsumeFaucetMetadataConfigNetwork]
             },
             PricedNote::Standard(StandardNote::ALLOWLIST_CONFIG) => {
                 &[ExecutionBenchmark::ConsumeAllowlistConfigNetwork]
@@ -449,7 +449,7 @@ mod tests {
     #[case::mint(PricedNote::Standard(StandardNote::MINT))]
     #[case::burn(PricedNote::Standard(StandardNote::BURN))]
     #[case::faucet_policy_config(PricedNote::Standard(StandardNote::FAUCET_POLICY_CONFIG))]
-    #[case::fungible_faucet_config(PricedNote::Standard(StandardNote::FUNGIBLE_FAUCET_CONFIG))]
+    #[case::faucet_metadata_config(PricedNote::Standard(StandardNote::FAUCET_METADATA_CONFIG))]
     #[case::allowlist_config(PricedNote::Standard(StandardNote::ALLOWLIST_CONFIG))]
     #[case::blocklist_config(PricedNote::Standard(StandardNote::BLOCKLIST_CONFIG))]
     #[case::pause_config(PricedNote::Standard(StandardNote::PAUSE_CONFIG))]